makefile: fix dependencies to allow make -j
Also, get rid of repeated Make rules using a suffix rule.
Suffix rules are described in POSIX and should be compatible
with both BSD and GNU Make.

The command used to generate the dependencies was as follows:
    perl -lne '
        if (eof and keys %mod) {
            print $ARGV=~s/\.f90/.o/r, ": ",
                join " ", map "$_.o", keys %mod;
                %mod=()
        };
        undef $mod{$1} if /^\s*use\s*(\w+)/;
    ' *.f90
Except the result had to be manually tweaked to remove intrinsic modules
and rename the object file which wasn't named after the module.